+++ next.config.js
@@ -1,6 +1,3 @@
-const withTM = require("next-transpile-modules")([
-  "library-simplified-webpub-viewer",
-]);
 const path = require("path");
 
 const config = {
@@ -9,6 +6,7 @@ const config = {
   },
   webpack: (config) => {
     config.resolve.alias["~"] = path.resolve(__dirname);
+
     return config;
   },
   sassOptions: {
@@ -16,4 +14,4 @@ const config = {
   },
 };
 
-module.exports = withTM(config);
+module.exports = config;