+++ src/__tests__/AdvancedSearch.test.tsx
@@ -93,6 +93,7 @@ describe("Advanced Search submit", () => {
     const expectedQuery = {
       filter: "language:english,startYear:1990,endYear:1999,format:pdf",
       query: "keyword:cat,author:Nook,title:Handbook,subject:poetry",
+      readerVersion: "v2",
     };
     expect(mockPush).toHaveBeenCalledTimes(1);
     expect(mockPush).toHaveBeenCalledWith({
@@ -120,6 +121,7 @@ describe("Advanced Search submit", () => {
     const expectedQuery = {
       filter: "startYear:1990",
       query: "keyword:cat",
+      readerVersion: "v2",
     };
     expect(mockPush).toHaveBeenCalledTimes(1);
     expect(mockPush).toHaveBeenCalledWith({
@@ -147,6 +149,7 @@ describe("Advanced Search submit", () => {
     const expectedQuery = {
       filter: "endYear:1990",
       query: "author:Shakespeare",
+      readerVersion: "v2",
     };
     expect(mockPush).toHaveBeenCalledTimes(1);
     expect(mockPush).toHaveBeenCalledWith({