+++ src/__tests__/componentHelpers/FilterYears.tsx
@@ -43,6 +43,7 @@ export const FilterYearsTests = (
         query: {
           filter: "startYear:1990",
           query: "keyword:Animal Crossing",
+          readerVersion: "v2",
         },
       });
     });
@@ -62,6 +63,7 @@ export const FilterYearsTests = (
         query: {
           filter: "endYear:1990",
           query: "keyword:Animal Crossing",
+          readerVersion: "v2",
         },
       });
     });
@@ -86,6 +88,7 @@ export const FilterYearsTests = (
         query: {
           filter: "startYear:1990,endYear:2000",
           query: "keyword:Animal Crossing",
+          readerVersion: "v2",
         },
       });
     });