+++ src/components/ReaderLayout/ReaderLayout.tsx
@@ -3,21 +3,27 @@ import * as DS from "@nypl/design-system-react-components";
 import { breadcrumbTitles } from "~/src/constants/labels";
 import { ApiLink, LinkResult } from "~/src/types/LinkQuery";
 import IFrameReader from "../IFrameReader/IFrameReader";
-import WebpubViewer from "../WebpubViewer/WebpubViewer";
 import EditionCardUtils from "~/src/util/EditionCardUtils";
 import Layout from "../Layout/Layout";
 import * as gtag from "../../lib/Analytics";
-import { truncateStringOnWhitespace } from "~/src/util/Util";
+import { formatUrl, truncateStringOnWhitespace } from "~/src/util/Util";
 import { MAX_TITLE_LENGTH } from "~/src/constants/editioncard";
-
+import dynamic from "next/dynamic";
+import { useRouter } from "next/router";
+const WebReader = dynamic(() => import("@nypl/web-reader"), { ssr: false });
 //The NYPL wrapper that wraps the Reader pages.
-const ReaderLayout: React.FC<{ linkResult: LinkResult }> = (props) => {
+const ReaderLayout: React.FC<{ linkResult: LinkResult; proxyUrl: string }> = (
+  props
+) => {
+  const router = useRouter();
+  const origin = router.basePath;
   const link: ApiLink = props.linkResult.data;
-
+  const proxyUrl = props.proxyUrl;
+  const url = formatUrl(link.url);
   const edition = link.work.editions[0];
 
-  const isRead = link.flags.reader;
   const isEmbed = link.flags.embed;
+  const isRead = link.flags.reader;
 
   useEffect(() => {
     gtag.drbEvents("Read", `${link.work.title}`);
@@ -46,7 +52,13 @@ const ReaderLayout: React.FC<{ linkResult: LinkResult }> = (props) => {
           <IFrameReader url={link.url} />
         </Layout>
       )}
-      {isRead && <WebpubViewer url={link.url} />}
+      {isRead && (
+        <WebReader
+          webpubManifestUrl={url}
+          proxyUrl={proxyUrl}
+          pdfWorkerSrc={`${origin}/pdf-worker/pdf.worker.min.js`}
+        />
+      )}
     </>
   );
 };